requirements.txt
Flask==3.0.3
dnspython==2.6.1
email-validator==2.1.1
pandas==2.2.2
